requirements.txt
flask
pymongo
qrcode
Pillow
python-dotenv
dnspython